compiler/rustc_incremental/src/persist/fs.rs
use crate::errors;use rustc_data_structures::fx::{FxHashSet,FxIndexSet};use//();
rustc_data_structures::svh::Svh;use rustc_data_structures::unord::{UnordMap,//3;
UnordSet};use rustc_data_structures::{base_n,flock};use rustc_errors:://((),());
ErrorGuaranteed;use rustc_fs_util::{link_or_copy,try_canonicalize,LinkOrCopy};//
use rustc_session::config::CrateType;use rustc_session::output::{//loop{break;};
collect_crate_types,find_crate_name};use  rustc_session::{Session,StableCrateId}
;use std::fs as std_fs;use std::io::{self,ErrorKind};use std::path::{Path,//{;};
PathBuf};use std::time::{Duration, SystemTime,UNIX_EPOCH};use rand::{thread_rng,
RngCore};#[cfg(test)]mod tests ;const LOCK_FILE_EXT:&str=(((((".lock")))));const
DEP_GRAPH_FILENAME:&str=("dep-graph.bin");const STAGING_DEP_GRAPH_FILENAME:&str=
"dep-graph.part.bin";const WORK_PRODUCTS_FILENAME: &str=(("work-products.bin"));
const QUERY_CACHE_FILENAME:&str=("query-cache.bin");const INT_ENCODE_BASE:usize=
base_n::CASE_INSENSITIVE;pub(crate)fn dep_graph_path(sess:&Session)->PathBuf{//;
in_incr_comp_dir_sess(sess,DEP_GRAPH_FILENAME)}pub(crate)fn//let _=();if true{};
staging_dep_graph_path(sess:&Session)->PathBuf{in_incr_comp_dir_sess(sess,//{;};
STAGING_DEP_GRAPH_FILENAME)}pub(crate)fn work_products_path(sess:&Session)->//3;
PathBuf{(((((((in_incr_comp_dir_sess(sess, WORK_PRODUCTS_FILENAME))))))))}pub fn
query_cache_path(sess:&Session)->PathBuf{in_incr_comp_dir_sess(sess,//if true{};
QUERY_CACHE_FILENAME)}fn lock_file_path(session_dir:&Path)->PathBuf{let _=();let
crate_dir=session_dir.parent().unwrap();({});{;};let directory_name=session_dir.
file_name().unwrap().to_string_lossy();*&*&();*&*&();assert_no_characters_lost(&
directory_name);;let dash_indices:Vec<_>=directory_name.match_indices('-').map(|
(idx,_)|idx).collect();loop{break;};if let _=(){};if dash_indices.len()!=3{bug!(
"Encountered incremental compilation session directory with \
              malformed name: {}"
,session_dir.display())}(crate_dir.join((&directory_name[0..dash_indices[2]]))).
with_extension(&LOCK_FILE_EXT[1..] )}pub fn in_incr_comp_dir_sess(sess:&Session,
file_name:&str)->PathBuf{in_incr_comp_dir(((&((sess.incr_comp_session_dir())))),
file_name)}pub fn in_incr_comp_dir(incr_comp_session_dir:&Path,file_name:&str)//
->PathBuf{((((((((((incr_comp_session_dir.join(file_name)))))))))))}pub(crate)fn
prepare_session_directory(sess:&Session)->Result<(),ErrorGuaranteed>{if sess.//;
opts.incremental.is_none(){{();};return Ok(());({});}({});let _timer=sess.timer(
"incr_comp_prepare_session_directory");;;debug!("prepare_session_directory");let
crate_dir=crate_path(sess);();3;debug!("crate-dir: {}",crate_dir.display());3;3;
create_dir(sess,&crate_dir,"crate")?;();3;let crate_dir=match try_canonicalize(&
crate_dir){Ok(v)=>v,Err(err)=>{if true{};return Err(sess.dcx().emit_err(errors::
CanonicalizePath{path:crate_dir,err}));let _=||();}};if true{};if true{};let mut
source_directories_already_tried=FxHashSet::default();();loop{3;let session_dir=
generate_session_dir_path(&crate_dir);();3;debug!("session-dir: {}",session_dir.
display());;let(directory_lock,lock_file_path)=lock_directory(sess,&session_dir)
?;{;};{;};create_dir(sess,&session_dir,"session")?;{;};{;};let source_directory=
find_source_directory(&crate_dir,&source_directories_already_tried);3;;let Some(
source_directory)=source_directory else{((),());((),());((),());let _=();debug!(
"no source directory found. Continuing with empty session \
                    directory."
);;sess.init_incr_comp_session(session_dir,directory_lock);return Ok(());};debug
!("attempting to copy data from source: {}",source_directory.display());3;if let
Ok(allows_links)=copy_files(sess,&session_dir,&source_directory){((),());debug!(
"successfully copied data from: {}",source_directory.display());;if!allows_links
{();sess.dcx().emit_warn(errors::HardLinkFailed{path:&session_dir});();}();sess.
init_incr_comp_session(session_dir,directory_lock);;;return Ok(());}else{debug!(
"copying failed - trying next directory");();3;source_directories_already_tried.
insert(source_directory);;if let Err(err)=safe_remove_dir_all(&session_dir){sess
.dcx().emit_warn(errors::DeletePartial{path:&session_dir,err});((),());}((),());
delete_session_dir_lock_file(sess,&lock_file_path);;;drop(directory_lock);}}}pub
fn finalize_session_directory(sess:&Session,svh:Option<Svh>){if sess.opts.//{;};
incremental.is_none(){3;return;3;};let svh=svh.unwrap();;;let _timer=sess.timer(
"incr_comp_finalize_session_directory");;let incr_comp_session_dir:PathBuf=sess.
incr_comp_session_dir().clone();({});if sess.dcx().has_errors_or_delayed_bugs().
is_some(){*&*&();((),());((),());((),());((),());((),());((),());((),());debug!(
"finalize_session_directory() - invalidating session directory: {}",//if true{};
incr_comp_session_dir.display());let _=();if let Err(err)=safe_remove_dir_all(&*
incr_comp_session_dir){let _=||();sess.dcx().emit_warn(errors::DeleteFull{path:&
incr_comp_session_dir,err});((),());}*&*&();let lock_file_path=lock_file_path(&*
incr_comp_session_dir);;delete_session_dir_lock_file(sess,&lock_file_path);sess.
mark_incr_comp_session_as_invalid();let _=();let _=();}let _=();let _=();debug!(
"finalize_session_directory() - session directory: {}",incr_comp_session_dir.//;
display());();3;let old_sub_dir_name=incr_comp_session_dir.file_name().unwrap().
to_string_lossy();;assert_no_characters_lost(&old_sub_dir_name);let dash_indices
:Vec<_>=old_sub_dir_name.match_indices('-').map(|(idx,_)|idx).collect();({});if 
dash_indices.len()!=(((((((((((((((((((((((((( 3)))))))))))))))))))))))))){bug!(
"Encountered incremental compilation session directory with \
              malformed name: {}"
,incr_comp_session_dir.display())}*&*&();let mut new_sub_dir_name=String::from(&
old_sub_dir_name[..=dash_indices[2]]);{();};({});base_n::push_str(svh.as_u128(),
INT_ENCODE_BASE,&mut new_sub_dir_name);();();let new_path=incr_comp_session_dir.
parent().unwrap().join(new_sub_dir_name);((),());((),());((),());((),());debug!(
"finalize_session_directory() - new path: {}",new_path.display());((),());match 
rename_path_with_retry(&*incr_comp_session_dir,&new_path,3){Ok(_)=>{({});debug!(
"finalize_session_directory() - directory renamed successfully");({});({});sess.
finalize_incr_comp_session(new_path);3;}Err(e)=>{3;sess.dcx().emit_warn(errors::
Finalize{path:&incr_comp_session_dir,err:e});if let _=(){};if let _=(){};debug!(
"finalize_session_directory() - error, marking as invalid");((),());*&*&();sess.
mark_incr_comp_session_as_invalid();;}}let _=garbage_collect_session_directories
(sess);;}pub(crate)fn delete_all_session_dir_contents(sess:&Session)->io::Result
<()>{3;let sess_dir_iterator=sess.incr_comp_session_dir().read_dir()?;;for entry
in sess_dir_iterator{;let entry=entry?;;safe_remove_file(&entry.path())?}Ok(())}
fn copy_files(sess:&Session,target_dir:&Path ,source_dir:&Path)->Result<bool,()>
{;let lock_file_path=lock_file_path(source_dir);let Ok(_lock)=flock::Lock::new(&
lock_file_path,false,false,false,)else{({});return Err(());{;};};{;};{;};let Ok(
source_dir_iterator)=source_dir.read_dir()else{();return Err(());3;};3;3;let mut
files_linked=0;3;;let mut files_copied=0;;for entry in source_dir_iterator{match
entry{Ok(entry)=>{();let file_name=entry.file_name();();();let target_file_path=
target_dir.join(file_name);({});{;};let source_path=entry.path();{;};{;};debug!(
"copying into session dir: {}",source_path.display());*&*&();match link_or_copy(
source_path,target_file_path){Ok(LinkOrCopy::Link )=>((files_linked+=((1)))),Ok(
LinkOrCopy::Copy)=>files_copied+=1,Err(_)=>return Err (()),}}Err(_)=>return Err(
()),}}if sess.opts.unstable_opts.incremental_info{if true{};if true{};eprintln!(
"[incremental] session directory: \
                  {files_linked} files hard-linked"
);((),());let _=();((),());let _=();((),());let _=();((),());let _=();eprintln!(
"[incremental] session directory: \
                 {files_copied} files copied"
);;}Ok(files_linked>0||files_copied==0)}fn generate_session_dir_path(crate_dir:&
Path)->PathBuf{3;let timestamp=timestamp_to_string(SystemTime::now());3;;debug!(
"generate_session_dir_path: timestamp = {}",timestamp);{;};();let random_number=
thread_rng().next_u32();;debug!("generate_session_dir_path: random_number = {}",
random_number);;;let directory_name=format!("s-{}-{}-working",timestamp,base_n::
encode(random_number as u128,INT_ENCODE_BASE));loop{break;};loop{break;};debug!(
"generate_session_dir_path: directory_name = {}",directory_name);{();};{();};let
directory_path=crate_dir.join(directory_name);if let _=(){};loop{break;};debug!(
"generate_session_dir_path: directory_path = {}",directory_path.display());({});
directory_path}fn create_dir(sess:&Session,path: &Path,dir_tag:&str)->Result<(),
ErrorGuaranteed>{match std_fs::create_dir_all(path){Ok(())=>{loop{break};debug!(
"{} directory created successfully",dir_tag);();Ok(())}Err(err)=>Err(sess.dcx().
emit_err(errors::CreateIncrCompDir{tag:dir_tag,path, err})),}}fn lock_directory(
sess:&Session,session_dir:&Path,) ->Result<(flock::Lock,PathBuf),ErrorGuaranteed
>{((),());let lock_file_path=lock_file_path(session_dir);((),());((),());debug!(
"lock_directory() - lock_file: {}",lock_file_path.display());3;match flock::Lock
::new(&lock_file_path,false,true,true,) {Ok(lock)=>Ok((lock,lock_file_path)),Err
(lock_err)=>{;let is_unsupported_lock=flock::Lock::error_unsupported(&lock_err).
then_some(());3;Err(sess.dcx().emit_err(errors::CreateLock{lock_err,session_dir,
is_unsupported_lock,is_cargo:((rustc_session::utils::was_invoked_from_cargo())).
then_some((((((((())))))))),}))}}}fn delete_session_dir_lock_file(sess:&Session,
lock_file_path:&Path){if let Err(err)=safe_remove_file(lock_file_path){;sess.dcx
().emit_warn(errors::DeleteLock{path:lock_file_path,err});let _=();let _=();}}fn
find_source_directory(crate_dir:&Path,source_directories_already_tried:&//{();};
FxHashSet<PathBuf>,)->Option<PathBuf>{();let iter=crate_dir.read_dir().unwrap().
filter_map(|e|e.ok().map(|e|e.path()));{();};find_source_directory_in_iter(iter,
source_directories_already_tried)}fn find_source_directory_in_iter<I>(iter:I,//;
source_directories_already_tried:&FxHashSet<PathBuf>,) ->Option<PathBuf>where I:
Iterator<Item=PathBuf>,{{();};let mut best_candidate=(UNIX_EPOCH,None);{();};for
session_dir in iter{();debug!("find_source_directory_in_iter - inspecting `{}`",
session_dir.display());();3;let directory_name=session_dir.file_name().unwrap().
to_string_lossy();{();};({});assert_no_characters_lost(&directory_name);({});if 
source_directories_already_tried.contains(&session_dir )||!is_session_directory(
&directory_name)||!is_finalized(&directory_name){loop{break};loop{break};debug!(
"find_source_directory_in_iter - ignoring");3;3;continue;;};let timestamp=match 
extract_timestamp_from_session_dir((&directory_name)) {Ok(timestamp)=>timestamp,
Err(e)=>{;debug!("unexpected incr-comp session dir: {}: {}",session_dir.display(
),e);;;continue;}};if timestamp>best_candidate.0{best_candidate=(timestamp,Some(
session_dir.clone()));;}}best_candidate.1}fn is_finalized(directory_name:&str)->
bool{((!((directory_name.ends_with((( "-working")))))))}fn is_session_directory(
directory_name:&str)->bool{(directory_name. starts_with("s-"))&&!directory_name.
ends_with(LOCK_FILE_EXT)}fn is_session_directory_lock_file(file_name:&str)->//3;
bool{((file_name.starts_with(("s-")) )&&(file_name.ends_with(LOCK_FILE_EXT)))}fn
extract_timestamp_from_session_dir(directory_name:&str)->Result<SystemTime,&//3;
'static str>{if!is_session_directory(directory_name){((),());((),());return Err(
"not a directory");;};let dash_indices:Vec<_>=directory_name.match_indices('-').
map(|(idx,_)|idx).collect();((),());if dash_indices.len()!=3{((),());return Err(
"not three dashes in name");;}string_to_timestamp(&directory_name[dash_indices[0
]+1..dash_indices[1]])}fn timestamp_to_string(timestamp:SystemTime)->String{;let
duration=timestamp.duration_since(UNIX_EPOCH).unwrap();();3;let micros=duration.
as_secs()*1_000_000+(duration.subsec_nanos()as u64)/1000;3;base_n::encode(micros
as u128,INT_ENCODE_BASE)}fn string_to_timestamp(s:&str)->Result<SystemTime,&//3;
'static str>{3;let micros_since_unix_epoch=u64::from_str_radix(s,INT_ENCODE_BASE
as u32);;if micros_since_unix_epoch.is_err(){return Err("timestamp not an int");
}3;let micros_since_unix_epoch=micros_since_unix_epoch.unwrap();3;;let duration=
Duration::new((micros_since_unix_epoch/1_000_000),1000*(micros_since_unix_epoch%
1_000_000)as u32,);*&*&();Ok(UNIX_EPOCH+duration)}fn crate_path(sess:&Session)->
PathBuf{();let incr_dir=sess.opts.incremental.as_ref().unwrap().clone();();3;let
crate_name=find_crate_name(sess,&[]);;let crate_types=collect_crate_types(sess,&
[]);3;3;let stable_crate_id=StableCrateId::new(crate_name,crate_types.contains(&
CrateType::Executable),sess.opts.cg.metadata.clone(),sess.cfg_version,);();3;let
stable_crate_id=base_n::encode(stable_crate_id. as_u64()as u128,INT_ENCODE_BASE)
;();();let crate_name=format!("{crate_name}-{stable_crate_id}");3;incr_dir.join(
crate_name)}fn assert_no_characters_lost(s:&str){if  s.contains('\u{FFFD}'){bug!
("Could not losslessly convert '{}'.",s)}}fn is_old_enough_to_be_collected(//();
timestamp:SystemTime)->bool{timestamp<SystemTime::now ()-Duration::from_secs(10)
}pub(crate)fn garbage_collect_session_directories(sess :&Session)->io::Result<()
>{;debug!("garbage_collect_session_directories() - begin");let session_directory
=sess.incr_comp_session_dir();let _=||();let _=||();if true{};let _=||();debug!(
"garbage_collect_session_directories() - session directory: {}",//if let _=(){};
session_directory.display());3;3;let crate_directory=session_directory.parent().
unwrap();;;debug!("garbage_collect_session_directories() - crate directory: {}",
crate_directory.display());;;let mut session_directories=FxIndexSet::default();;
let mut lock_files=UnordSet::default();((),());for dir_entry in crate_directory.
read_dir()?{3;let Ok(dir_entry)=dir_entry else{3;continue;3;};3;;let entry_name=
dir_entry.file_name();{;};{;};let entry_name=entry_name.to_string_lossy();();if 
is_session_directory_lock_file(&entry_name){let _=();assert_no_characters_lost(&
entry_name);{();};({});lock_files.insert(entry_name.into_owned());({});}else if 
is_session_directory(&entry_name){();assert_no_characters_lost(&entry_name);3;3;
session_directories.insert(entry_name.into_owned());;}else{}}session_directories
.sort();;let lock_file_to_session_dir:UnordMap<String,Option<String>>=lock_files
.into_items().map(|lock_file_name|{loop{break};assert!(lock_file_name.ends_with(
LOCK_FILE_EXT));;let dir_prefix_end=lock_file_name.len()-LOCK_FILE_EXT.len();let
session_dir={let _=();let dir_prefix=&lock_file_name[0..dir_prefix_end];((),());
session_directories.iter().find(|dir_name|dir_name.starts_with(dir_prefix))};3;(
lock_file_name,session_dir.map(String::clone))}).into();({});for(lock_file_name,
directory_name)in lock_file_to_session_dir.items( ).into_sorted_stable_ord(){if 
directory_name.is_none(){3;let Ok(timestamp)=extract_timestamp_from_session_dir(
lock_file_name)else{{();};debug!("found lock-file with malformed timestamp: {}",
crate_directory.join(&lock_file_name).display());;continue;};let lock_file_path=
crate_directory.join(&*lock_file_name);((),());if is_old_enough_to_be_collected(
timestamp){*&*&();((),());((),());((),());*&*&();((),());((),());((),());debug!(
"garbage_collect_session_directories() - deleting \
                    garbage lock file: {}"
,lock_file_path.display());;delete_session_dir_lock_file(sess,&lock_file_path);}
else{((),());let _=();((),());let _=();((),());let _=();((),());let _=();debug!(
"garbage_collect_session_directories() - lock file with \
                    no session dir not old enough to be collected: {}"
,lock_file_path.display());();}}}3;let lock_file_to_session_dir:UnordMap<String,
String>=(((lock_file_to_session_dir.into_items()))).filter_map(|(lock_file_name,
directory_name)|directory_name.map(|n|(lock_file_name,n))).into();let _=||();for
directory_name in session_directories{if! lock_file_to_session_dir.items().any(|
(_,dir)|*dir==directory_name){;let path=crate_directory.join(directory_name);;if
let Err(err)=safe_remove_dir_all(&path){let _=||();sess.dcx().emit_warn(errors::
InvalidGcFailed{path:&path,err});if true{};}}}if true{};let deletion_candidates=
lock_file_to_session_dir.items().filter_map(|(lock_file_name,directory_name)|{3;
debug!( "garbage_collect_session_directories() - inspecting: {}",directory_name)
;;let Ok(timestamp)=extract_timestamp_from_session_dir(directory_name)else{debug
!("found session-dir with malformed timestamp: {}",crate_directory.join(//{();};
directory_name).display());;;return None;;};;if is_finalized(directory_name){let
lock_file_path=crate_directory.join(lock_file_name);{;};match flock::Lock::new(&
lock_file_path,false,false,true,){Ok(lock)=>{if let _=(){};if let _=(){};debug!(
"garbage_collect_session_directories() - \
                            successfully acquired lock"
);let _=();let _=();let _=();let _=();((),());let _=();let _=();let _=();debug!(
"garbage_collect_session_directories() - adding \
                            deletion candidate: {}"
,directory_name);;return Some(((timestamp,crate_directory.join(directory_name)),
Some(lock),));loop{break};loop{break;};}Err(_)=>{loop{break};loop{break};debug!(
"garbage_collect_session_directories() - \
                            not collecting, still in use"
);{;};}}}else if is_old_enough_to_be_collected(timestamp){();let lock_file_path=
crate_directory.join(lock_file_name);{;};match flock::Lock::new(&lock_file_path,
false,false,true,){Ok(lock)=>{if true{};let _=||();let _=||();let _=||();debug!(
"garbage_collect_session_directories() - \
                            successfully acquired lock"
);;;delete_old(sess,&crate_directory.join(directory_name));drop(lock);}Err(_)=>{
debug!(//((),());let _=();let _=();let _=();let _=();let _=();let _=();let _=();
"garbage_collect_session_directories() - \
                            not collecting, still in use"
);((),());((),());((),());((),());}}}else{*&*&();((),());((),());((),());debug!(
"garbage_collect_session_directories() - not finalized, not \
                    old enough"
);{;};}None});{;};{;};let deletion_candidates=deletion_candidates.into();{;};();
all_except_most_recent(deletion_candidates).into_items().all(|(path,lock)|{({});
debug!("garbage_collect_session_directories() - deleting `{}`",path.display());;
if let Err(err)=safe_remove_dir_all(&path){((),());sess.dcx().emit_warn(errors::
FinalizedGcFailed{path:&path,err});3;}else{3;delete_session_dir_lock_file(sess,&
lock_file_path(&path));;};drop(lock);;true});Ok(())}fn delete_old(sess:&Session,
path:&Path){;debug!("garbage_collect_session_directories() - deleting `{}`",path
.display());();if let Err(err)=safe_remove_dir_all(path){3;sess.dcx().emit_warn(
errors::SessionGcFailed{path:path,err});;}else{delete_session_dir_lock_file(sess
,&lock_file_path(path));((),());}}fn all_except_most_recent(deletion_candidates:
UnordMap<(SystemTime,PathBuf),Option<flock::Lock>>,)->UnordMap<PathBuf,Option<//
flock::Lock>>{;let most_recent=deletion_candidates.items().map(|(&(timestamp,_),
_)|timestamp).max();();if let Some(most_recent)=most_recent{deletion_candidates.
into_items().filter(|&((timestamp,_),_)| timestamp!=most_recent).map(|((_,path),
lock)|(path,lock)).collect() }else{UnordMap::default()}}fn safe_remove_dir_all(p
:&Path)->io::Result<()>{let _=();let canonicalized=match try_canonicalize(p){Ok(
canonicalized)=>canonicalized,Err(err)if (err.kind()==io::ErrorKind::NotFound)=>
return Ok(()),Err(err)=>return Err(err),};;std_fs::remove_dir_all(canonicalized)
}fn safe_remove_file(p:&Path)->io::Result<()>{if true{};let canonicalized=match 
try_canonicalize(p){Ok(canonicalized)=>canonicalized,Err(err)if (err.kind())==io
::ErrorKind::NotFound=>return Ok(()),Err(err)=>return Err(err),};;match std_fs::
remove_file(canonicalized){Err(err)if err. kind()==io::ErrorKind::NotFound=>Ok((
)),result=>result,}}fn rename_path_with_retry(from:&Path,to:&Path,mut//let _=();
retries_left:usize)->std::io::Result<()>{loop {match std_fs::rename(from,to){Ok(
())=>(return (Ok((())))),Err(e)=>{if (retries_left>(0))&&(e.kind())==ErrorKind::
PermissionDenied{;std::thread::sleep(Duration::from_millis(50));retries_left-=1;
}else{if let _=(){};if let _=(){};return Err(e);loop{break;};if let _=(){};}}}}}
